* Palettes/3Containers/GormNSOutlineView.m         * GormCore/GormMatrixEditor.m         * GormCore/GormClassEditor.m           some int -> NSInteger and float -> CGFloat transitions           spotted by libobjc2 runtime in debug mode
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@36656 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassEditor.m
@@ -1191,7 +1191,7 @@ objectValueForTableColumn: (NSTableColumn *)aTableColumn
   [gov setNeedsDisplay: YES];
 }
 
-- (int) outlineView: (NSOutlineView *)anOutlineView 
+- (NSInteger) outlineView: (NSOutlineView *)anOutlineView 
 numberOfChildrenOfItem: (id)item
 {
   NSArray *subclasses = [classManager subClassesOf: item];
@@ -1213,7 +1213,7 @@ numberOfChildrenOfItem: (id)item
 }
 
 - (id) outlineView: (NSOutlineView *)anOutlineView 
-	     child: (int)index
+	     child: (NSInteger)index
 	    ofItem: (id)item
 {
   NSArray *subclasses = [classManager subClassesOf: item];
